Refine referral rewards modal
